# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Food Ordering API
  description: API for managing users, food items, and orders.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local Development Server
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        201:
          description: Registration requested successfully
        400:
          description: User already exists
  /auth/login:
    post:
      tags: [Auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        200:
          description: Login successful
        401:
          description: Invalid credentials
        403:
          description: User not approved
...